Add support for @JsExport annotation
Annotating classes or methods with this annotation causes the symbol to
be exported to other JavaScript code outside the control of j2cl,
preventing them from being removed due to dead code pruning.

When a class is annotated, the result is equivalent to calling
goog.exportSymbol with the qualified JavaScript name of the class.

When a method is annotated, the result is equivalent to annotating
@export in the method's JsDoc comment.

This change only covers support for the jdt frontend and the closure
backend.

The @JsExport annotation itself needs to be defined in a shared library.
Currently, we expect it to have no parameters:

@Retention(RetentionPolicy.RUNTIME)
@Target({ElementType.METHOD, ElementType.TYPE})
public @interface JsExport {
}