Support native-stub generation for external JavaScript objects
There may be JavaScript objects created outside of j2cl/closure that we
want to access from Java code.

For these objects, we create Java interfaces that are annotated with a
certain annotation that will mark these interfaces as "JavaScript
external". In order to reference these classes, we have to generate
"native stub" files so jscomp/closure doesn't get confused with the
type information.

The annotation used is either @JsImport or @DumboService (both declared
elsewhere). JsImport is of the following form:

@Retention(RetentionPolicy.RUNTIME)
@Target(ElementType.TYPE)
@Documented
public @interface JsImport {
}

Moreover, we have to mark the methods of these imported types with
"@export", and generate a corresponding reference in an "@externs" file
such that the names are guarded against renaming during code
optimization.

This change adds specific support for the @DumboService annotation to
create such files, which allows us to reuse existing dependencies
without requiring them to declare a new annotation. While we're at it,
we also register the corresponding generated JavaScript classname with
Dumbo, so we can resolve the corresponding Dumbo service JavaScript
object from Java by specifying its RPC name.

This change only covers support for the jdt frontend and the closure
backend.