Support vanilla Java implementation in a JsType isNative=true class
Sometimes, we want to have a class that has a vanilla Java
implementation, but in the context of j2cl/JavaScript it should have a
JavaScript-native implementation.

Right now, specifying a class with @JsType(isNative=true) prevents us
from doing that.

The official recommendation is to resort to "super-sourcing", which
means using different source code for each target environment
(Java/JavaScript). Unfortunately, this makes the development of shared
libraries unnecessarily difficult.

This change adds support for methods that have a new
@JsImplementationProvidedSeparately annotation. Such methods have their
method body for vanilla Java (as usual), but when transpiling to
JavaScript, the JavaScript code will be taken from where the rest of the
class (i.e., the other "native" methods) are defined.

This change only covers support for the jdt frontend and the closure
backend.

The annotation is defined elsewhere, and expected to be as follows:

@Retention(RetentionPolicy.RUNTIME)
@Target({ElementType.METHOD})
@Documented
public @interface JsImplementationProvidedSeparately {
}

Also see
https://github.com/google/j2cl/issues/221