Add support for @JsIgnoreType
Sometimes, sharing the same source folder between code that should be
transpiled and code that shouldn't is desirable. Currently, this is not
possible without heavy use of @JsIgnore, which is not available at the
type level.

This change adds support for the @JsIgnoreType annotation, which is
meant to declare code that should be ignored by j2cl.

This change only covers support for the jdt frontend and the closure
backend.

The annotation is defined elsewhere, and expected to be as follows:

@Retention(RetentionPolicy.RUNTIME)
@Target({ElementType.TYPE})
@Documented
public @interface JsIgnoreType {
}