Add support for @JsServiceProvider
Currently, j2cl lacks the ability to use the Service Provider Interface
(via java.util.ServiceLoader) from JavaScript.

This change adds this functionality, assuming a corresponding change to
jre.js (in a separate commit).

Any class annotated with @JsServiceProvider can declare its service
implementations. The class MUST implement these interfaces, and provide
a no-arg constructor with no declared exceptions.

The instance will be instantiated upon demand, but code is initialized
along with types annotated with @JsEntryPoint, so the services might not
be available if there is a dependency cycle upon calling a particular
static initializer).

This change only covers support for the jdt frontend and the closure
backend.

The annotation is defined elsewhere, and expected to be as follows:

@Retention(RetentionPolicy.RUNTIME)
@Target({ElementType.TYPE})
@Documented
public @interface JsServiceProvider {
  /**
   * The services implemented by the annotated class.
   *
   * @return The list of services.
   */
  Class<?>[] value() default {};
}