Add support for @JsPatched
Sometimes, it's necessary to patch or override the $impl code generated
by j2cl with some custom JavaScript, especially when
@JsImplementationProvidedSeparately is being used. Right now, there's no
easy way to declare this in Java.

By introducing support for a @JsPatched annotation, j2cl emits
JavaScript header files that instead of $impl source a $impl$patched
module, which is provided separately. That $impl$patched module may
either source and modify the created $impl file, or replace it entirely
with some other code.

This change only covers support for the jdt frontend and the closure
backend.

The annotation is defined elsewhere, and expected to be as follows:

@Retention(RetentionPolicy.RUNTIME)
@Target({ElementType.TYPE})
@Documented
public @interface JsPatched {
}