transpiler: CompilationBuilder: Fix compiler warning ("unlikely")
In Eclipse, we get the following warning
Unlikely argument type Symbol for get(Object) on a
Map<VariableElement,Variable>

The problem is that we declare a new variable casting an existing one to
a different type, but then later use the original variable in a context
that is meant for the new type.

Using the new type intead of the old type resolves the warning.